refactor(studio): extract useDomSelection from useDomEditSession
